+++ src/render/draw_symbol.test.ts
@@ -14,6 +14,7 @@ import {Transform} from '../geo/transform';
 import type {EvaluationParameters} from '../style/evaluation_parameters';
 import type {SymbolLayerSpecification} from '@maplibre/maplibre-gl-style-spec';
 import {Style} from '../style/style';
+import {MercatorProjection} from '../geo/projection/mercator';
 
 jest.mock('./painter');
 jest.mock('./program');
@@ -45,7 +46,8 @@ describe('drawSymbol', () => {
         painterMock.transform = {pitch: 0, labelPlaneMatrix: mat4.create()} as any as Transform;
         painterMock.options = {} as any;
         painterMock.style = {
-            map: {}
+            map: {},
+            projection: new MercatorProjection()
         } as any as Style;
 
         const layerSpec = {
@@ -62,7 +64,7 @@ describe('drawSymbol', () => {
 
         const tileId = new OverscaledTileID(1, 0, 1, 0, 0);
         tileId.posMatrix = mat4.create();
-        const programMock = new Program(null, null, null, null, null, null);
+        const programMock = new Program(null, null, null, null, null, null, null, null);
         (painterMock.useProgram as jest.Mock).mockReturnValue(programMock);
         const bucketMock = new SymbolBucket(null);
         bucketMock.icon = {
@@ -124,7 +126,7 @@ describe('drawSymbol', () => {
 
         const tileId = new OverscaledTileID(1, 0, 1, 0, 0);
         tileId.posMatrix = mat4.create();
-        const programMock = new Program(null, null, null, null, null, null);
+        const programMock = new Program(null, null, null, null, null, null, null, null);
         (painterMock.useProgram as jest.Mock).mockReturnValue(programMock);
         const bucketMock = new SymbolBucket(null);
         bucketMock.icon = {
@@ -150,7 +152,8 @@ describe('drawSymbol', () => {
         (sourceCacheMock.getTile as jest.Mock).mockReturnValue(tile);
         sourceCacheMock.map = {showCollisionBoxes: false} as any as Map;
         painterMock.style = {
-            map: {}
+            map: {},
+            projection: new MercatorProjection()
         } as any as Style;
 
         const spy = jest.spyOn(symbolProjection, 'updateLineLabels');
@@ -172,7 +175,7 @@ describe('drawSymbol', () => {
         painterMock.transform = {pitch: 0, labelPlaneMatrix: mat4.create()} as any as Transform;
         painterMock.options = {} as any;
         painterMock.style = {
-            map: {}
+            projection: new MercatorProjection()
         } as any as Style;
 
         const layerSpec = {
@@ -189,7 +192,7 @@ describe('drawSymbol', () => {
 
         const tileId = new OverscaledTileID(1, 0, 1, 0, 0);
         tileId.posMatrix = mat4.create();
-        const programMock = new Program(null, null, null, null, null, null);
+        const programMock = new Program(null, null, null, null, null, null, null, null);
         (painterMock.useProgram as jest.Mock).mockReturnValue(programMock);
         const bucketMock = new SymbolBucket(null);
         bucketMock.icon = {