+++ src/render/terrain.ts
@@ -384,7 +384,7 @@ export class Terrain {
             indexArray.emplaceBack(x + y, meshSize + x + y + 1, meshSize + x + y + 2);
             indexArray.emplaceBack(x + y, meshSize + x + y + 2, x + y + 1);
         }
-        // add an extra frame around the mesh to avoid stiching on tile boundaries with different zoomlevels
+        // add an extra frame around the mesh to avoid stitching on tile boundaries with different zoomlevels
         // first code-block is for top-bottom frame and second for left-right frame
         const offsetTop = vertexArray.length, offsetBottom = offsetTop + (meshSize + 1) * 2;
         for (const y of [0, 1]) for (let x = 0; x <= meshSize; x++) for (const z of [0, 1])
@@ -413,7 +413,7 @@ export class Terrain {
     }
 
     /**
-     * Calculates a height of the frame around the terrain-mesh to avoid stiching between
+     * Calculates a height of the frame around the terrain-mesh to avoid stitching between
      * tile boundaries in different zoomlevels.
      * @param zoom - current zoomlevel
      * @returns the elevation delta in meters