+++ src/shaders/clipping_mask.vertex.glsl
@@ -1,7 +1,5 @@
 in vec2 a_pos;
 
-uniform mat4 u_matrix;
-
 void main() {
-    gl_Position = u_matrix * vec4(a_pos, 0, 1);
+    gl_Position = projectTile(a_pos);
 }