+++ src/source/geojson_source.test.ts
@@ -6,6 +6,7 @@ import {LngLat} from '../geo/lng_lat';
 import {extend} from '../util/util';
 import {Dispatcher} from '../util/dispatcher';
 import {RequestManager} from '../util/request_manager';
+import {SubdivisionGranularitySetting} from '../render/subdivision_granularity_settings';
 import {ActorMessage, MessageType} from '../util/actor_messages';
 import {Actor} from '../util/actor';
 
@@ -389,7 +390,14 @@ describe('GeoJSONSource#update', () => {
         const source = new GeoJSONSource('id', {data: {}} as GeoJSONSourceOptions, mockDispatcher, undefined);
         source.map = {
             transform: {} as Transform,
-            getPixelRatio() { return 1; }
+            getPixelRatio() { return 1; },
+            style: {
+                projection: {
+                    get subdivisionGranularity() {
+                        return SubdivisionGranularitySetting.noSubdivision;
+                    }
+                }
+            }
         } as any;
 
         source.on('data', (e) => {