+++ src/source/vector_tile_source.test.ts
@@ -9,6 +9,7 @@ import fixturesSource from '../../test/unit/assets/source.json' with {type: 'jso
 import {getMockDispatcher, getWrapDispatcher, sleep, waitForMetadataEvent} from '../util/test/util';
 import {Map} from '../ui/map';
 import {WorkerTileParameters} from './worker_source';
+import {SubdivisionGranularitySetting} from '../render/subdivision_granularity_settings';
 import {ActorMessage, MessageType} from '../util/actor_messages';
 
 function createSource(options, transformCallback?, clearTiles = () => {}) {
@@ -17,8 +18,15 @@ function createSource(options, transformCallback?, clearTiles = () => {}) {
         transform: {showCollisionBoxes: false},
         _getMapId: () => 1,
         _requestManager: new RequestManager(transformCallback),
-        style: {sourceCaches: {id: {clearTiles}}},
-        getPixelRatio() { return 1; }
+        style: {
+            sourceCaches: {id: {clearTiles}},
+            projection: {
+                get subdivisionGranularity() {
+                    return SubdivisionGranularitySetting.noSubdivision;
+                }
+            }
+        },
+        getPixelRatio() { return 1; },
     } as any as Map);
 
     source.on('error', () => { }); // to prevent console log of errors