+++ test/bench/lib/tile_parser.ts
@@ -15,6 +15,7 @@ import type {OverscaledTileID} from '../../../src/source/tile_id';
 import type {TileJSON} from '../../../src/util/util';
 import type {Map} from '../../../src/ui/map';
 import type {IActor} from '../../../src/util/actor';
+import {SubdivisionGranularitySetting} from '../../../src/render/subdivision_granularity_settings';
 import {MessageType} from '../../../src/util/actor_messages';
 
 class StubMap extends Evented {
@@ -133,11 +134,12 @@ export default class TileParser {
             pixelRatio: 1,
             request: {url: ''},
             returnDependencies,
-            promoteId: undefined
+            promoteId: undefined,
+            subdivisionGranularity: SubdivisionGranularitySetting.noSubdivision
         });
 
         const vectorTile = new VT.VectorTile(new Protobuf(tile.buffer));
 
-        return workerTile.parse(vectorTile, this.layerIndex, [], this.actor);
+        return workerTile.parse(vectorTile, this.layerIndex, [], this.actor, SubdivisionGranularitySetting.noSubdivision);
     }
 }